requirements.txt
streamlit
pandas
plotly
openpyxl
streamlit
psycopg2-binary
pandas
bcrypt
python-dotenv
